requirements.txt
requests
contentful
python-dotenv